tx builder update and minting unit/demo
tx builder was overriding the Witness SKey which is ok if you run once, but since you will naturally need to rerun due to fee calculation, we needed to stop the override. Will need SimpleExec to turn minting test into an actual test. since our fee calculation is very close but not 1 to 1